backend/requirements.txt
# --- 基础工具 ---
requests==2.32.5
PyYAML==6.0.3
psutil==7.2.2
packaging==26.0

# --- 后端框架 (你需要加的) ---
fastapi
uvicorn
websockets
aiohttp

# --- 系统增强 ---
lxml==6.0.2  # 如果你需要解析一些复杂的 XML 数据（如某些老的源）